Adds carbon footprint component
